+++ apps/mobile/src/app/(home,search,profile,notifications,settings)/settings/defaults.tsx
@@ -1,26 +1,63 @@
 import { create } from 'mutative'
-import { ScrollView } from 'react-native'
+import { FlatList, ScrollView } from 'react-native'
 import ReorderableList, { reorderItems } from 'react-native-reorderable-list'
 import { createStyleSheet, useStyles } from 'react-native-unistyles'
 import { useTranslations } from 'use-intl'
 
+import { Switch } from '~/components/common/switch'
 import { Text } from '~/components/common/text'
 import { DraggableItem } from '~/components/defaults/draggable-item'
+import { SheetItem } from '~/components/sheets/item'
 import { useList } from '~/hooks/list'
+import { FeedTypeColors, FeedTypeIcons } from '~/lib/sort'
 import { useDefaults } from '~/stores/defaults'
+import { FeedType } from '~/types/sort'
 
 export default function Screen() {
   const t = useTranslations('screen.settings.defaults')
 
-  const { drawerSections, searchTabs, update } = useDefaults()
+  const { drawerSections, feedType, searchTabs, update } = useDefaults()
 
-  const { styles } = useStyles(stylesheet)
+  const { styles, theme } = useStyles(stylesheet)
 
   const listProps = useList()
 
   return (
     <ScrollView {...listProps}>
       <Text highContrast={false} mb="2" mt="3" mx="3" size="2" weight="medium">
+        {t('feedType.title')}
+      </Text>
+
+      <FlatList
+        data={FeedType}
+        renderItem={({ item }) => (
+          <SheetItem
+            icon={{
+              color: theme.colors[FeedTypeColors[item]].accent,
+              name: FeedTypeIcons[item],
+              size: theme.space[4],
+              type: 'icon',
+              weight: 'duotone',
+            }}
+            label={t(`feedType.${item}`)}
+            onPress={() => null}
+            right={
+              <Switch
+                onChange={() => {
+                  update({
+                    feedType: item,
+                  })
+                }}
+                value={item === feedType}
+              />
+            }
+          />
+        )}
+        scrollEnabled={false}
+        style={styles.list}
+      />
+
+      <Text highContrast={false} mb="2" mt="6" mx="3" size="2" weight="medium">
         {t('searchTabs.title')}
       </Text>